refactor: split performance returns payloads
