+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -13,7 +13,7 @@ This ledger tracks the enterprise refactor continuation from
 | RISK-REF-006 | Observability and operational diagnostics | Hardened | Request middleware emitted an interpolated text log, weakening production parsing and field-level sensitive-data proof. | Replaced it with a bounded structured `request_observation` event and tests proving query strings and bodies are excluded. Continue reviewing operator diagnostics and production threshold evidence. | `tests/unit/test_correlation_middleware.py`; `make observability-contract-validate`; observability tests |
 | RISK-REF-007 | Skills, guidance, documentation, and agent context | Hardened | Long-running refactor guidance did not require a recurring conscious review of agent effectiveness artifacts. | Added a five-slice/major-decision review cadence and durable register. Existing backend-delivery and review-ledger skills remain sufficient; repository context remains accurate. | `quality/agent_effectiveness_review.md`; `tests/unit/test_agent_effectiveness_review.py` |
 | RISK-REF-008 | Source module size regression | Signed Off | Current source modularization keeps the largest Python source file at 401 lines, but CI only reported size and could allow new monoliths. | Added an active 450-line source-file regression gate to local `check`/`ci` and all GitHub delivery lanes, with focused tests. | `make source-size-gate`; `tests/unit/test_source_size_gate.py`; workflow lint |
-| RISK-REF-009 | Downstream connection lifecycle | Hardened | Both downstream profiles configured pool limits, but each operation constructed and closed a new HTTP client, preventing cross-request connection reuse and leaving shutdown ownership implicit. Profile resolution also mixed environment parsing with HTTP-client construction and request execution. | Added application-lifespan ownership of reusable dependency-specific HTTP pools, explicit draining/close behavior, and compatibility-preserving injected/standalone paths. Extracted timeout, pool, keepalive, and async polling environment parsing into `downstream_profile_env.py`. Continue reviewing bounded retry posture separately because mutation retries require operation-specific idempotency guarantees. | `tests/unit/test_app_lifecycle.py`; downstream client tests; health tests; `python -m mypy --config-file mypy.ini` |
+| RISK-REF-009 | Downstream connection lifecycle | Hardened | Both downstream profiles configured pool limits, but each operation constructed and closed a new HTTP client, preventing cross-request connection reuse and leaving shutdown ownership implicit. Profile resolution also mixed environment parsing with HTTP-client construction and request execution; lotus-performance async returns polling mixed poll orchestration with payload validation and bounded async failure construction. | Added application-lifespan ownership of reusable dependency-specific HTTP pools, explicit draining/close behavior, and compatibility-preserving injected/standalone paths. Extracted timeout, pool, keepalive, and async polling environment parsing into `downstream_profile_env.py`; extracted lotus-performance async accepted/result payload validation and failure/timeout error construction into `performance_returns_payloads.py`. Continue reviewing bounded retry posture separately because mutation retries require operation-specific idempotency guarantees. | `tests/unit/test_app_lifecycle.py`; downstream client tests; lotus-performance client tests; health tests; `python -m mypy --config-file mypy.ini` |
 | RISK-REF-010 | Runtime configuration and downstream URL safety | Hardened | Core and performance URL resolution duplicated permissive handling; malformed or credential-bearing values failed late, and the required configuration guide was missing. | Added a shared safe URL resolver, negative tests, complete runtime configuration guide, expanded environment example, and security/runbook evidence. | `tests/unit/test_downstream_base_url.py`; `tests/unit/test_configuration_docs.py`; downstream client tests; `make security-audit` |
 | RISK-REF-011 | Enterprise bank runtime enforcement | Hardened | Documented bank mode required authorization and explicit settings, but enabling runtime enforcement did not require that posture and could start permissively. | Made enterprise runtime enforcement fail application construction with bounded issue codes unless the in-process bank posture is complete; kept local defaults unchanged. | `tests/unit/test_enterprise_readiness.py`; deployment-policy and configuration-doc tests; `make security-audit` |
 | RISK-REF-012 | Enterprise authorization module boundary | Hardened | `enterprise_readiness.py` mixed capability-rule parsing, write authorization, runtime validation, redaction, audit emission, policy metadata, payload limits, and middleware assembly, making the security boundary harder to inspect. | Extracted capability-rule parsing and write authorization into `enterprise_authorization.py`; extracted audit emission and sensitive metadata redaction into `enterprise_audit.py`; extracted policy-version lookup into `enterprise_policy.py`; preserved compatibility exports from `enterprise_readiness.py`. | `tests/unit/test_enterprise_readiness.py`; app factory/configuration/deployment-policy/security-evidence tests; `python -m mypy --config-file mypy.ini`; touched radon MI/CC checks |
